backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
beautifulsoup4==4.12.3
requests==2.32.3
pydantic==2.10.3
python-dotenv==1.0.1
api-analytics==1.2.7
supabase>=2.10.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
pytest-cov==6.0.0
